move constants file to root of web app
